refactor(ui-judge): split judge layers
