+++ .github/ui-judge.instructions.md
@@ -4,6 +4,8 @@ applyTo: "packages/genui/ui-judge/**/*"
 
 When extending `@lynx-js/ui-judge`, keep the public runtime API small and platform-specific. Playwright callers use `judgePage` and own page setup, navigation, viewport, cookies, route mocks, and authentication. Additional dimensions should remain internal unless they are intentionally added to the package exports.
 
+Keep `packages/genui/ui-judge/src/index.ts` as a public facade that only re-exports the supported API and public types. Put shared judge orchestration, option normalization, dimensions, and prompt construction under `src/core`, and keep platform-specific adapters such as Playwright and Kitten-Lynx/Android outside `core`, with low-level device or screenshot helpers under `src/platforms`.
+
 When adding Android support to `@lynx-js/ui-judge`, keep the public call shape close to `judgePage`: accept the `KittenLynxView` returned by `@lynx-js/kitten-lynx-test-infra`'s `newPage()` as `page`. Callers should own the Kitten-Lynx connection, navigation, and teardown lifecycle, while UI Judge creates the internal Midscene agent adapter. For Android scoring, pass `screenshotIncluded: true` without web-only DOM requirements, and return `page.url()` through the existing result `url` field.
 
 Keep Android-specific `@lynx-js/ui-judge` tests on Vitest rather than Playwright. Use a dedicated `test:android` script and let Playwright tests stay under `test:playwright`, so the Android emulator CI job can run UI Judge's Kitten-Lynx coverage without pulling in browser fixtures.
